move wrapper to interface
