+++ FWCore/PyDevParameterSet/src/PyBind11Wrapper.cc
@@ -1,4 +1,4 @@
-#include "PyBind11Wrapper.h"
+#include "FWCore/PyDevParameterSet/interface/PyBind11Wrapper.h"
 #include "FWCore/Utilities/interface/Exception.h"
 //#include <iostream>
 namespace edm {
